Questions can display with no header item
Added a condition for nesting at the first item level to check if the first item is a standalone question to be displayed or a conditional question with sub-questions. This allows the ability to display questions at the first level which are not stored within a "header" item above them.